+++ server/src/main/java/org/jetbrains/bsp/bazel/server/diagnostics/DiagnosticsService.kt
@@ -1,10 +1,13 @@
 package org.jetbrains.bsp.bazel.server.diagnostics
 
+import ch.epfl.scala.bsp4j.BuildTargetIdentifier
+import ch.epfl.scala.bsp4j.Diagnostic
 import ch.epfl.scala.bsp4j.PublishDiagnosticsParams
-import org.jetbrains.bsp.bazel.bazelrunner.BazelInfo
+import ch.epfl.scala.bsp4j.TextDocumentIdentifier
 import java.nio.file.Path
+import org.jetbrains.bsp.bazel.commons.BspCompileState
 
-class DiagnosticsService(workspaceRoot: Path) {
+class DiagnosticsService(workspaceRoot: Path, private val state: BspCompileState) {
 
     private val parser = DiagnosticsParser()
     private val mapper = DiagnosticBspMapper(workspaceRoot)
@@ -15,7 +18,27 @@ class DiagnosticsService(workspaceRoot: Path) {
         originId: String?
     ): List<PublishDiagnosticsParams> {
         val parsedDiagnostics = parser.parse(bazelOutput, targetLabel)
-        return mapper.createDiagnostics(parsedDiagnostics, originId)
+        val events = mapper.createDiagnostics(parsedDiagnostics, originId)
+        events
+            .groupBy { it.getBuildTarget().getUri() }
+            .forEach {
+                val buildTarget = it.key
+                val paramss = it.value
+                val docs = paramss.map { it.getTextDocument() }.toSet()
+                state.hasAnyProblems.put(buildTarget, docs)
+            }
+        return events
     }
 
+    fun clearFormerDiagnostics(targetLabel: String): List<PublishDiagnosticsParams> {
+        val id = targetLabel
+        val docs = state.hasAnyProblems[id]
+        if (docs != null && docs.isNotEmpty()) {
+            val diagnostics = emptyList<Diagnostic>()
+            state.hasAnyProblems.remove(id)
+            return docs.map { PublishDiagnosticsParams(it, BuildTargetIdentifier(id), diagnostics, true) }
+        } else {
+            return emptyList<PublishDiagnosticsParams>()
+        }
+    }
 }